Extract Email and Phone number validation to backend-foundation, remove shared ValidationUtils
