+++ account-management/Application/Tenants/Commands/CreateTenant.cs
@@ -3,6 +3,7 @@ using JetBrains.Annotations;
 using MediatR;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.Foundation.DomainModeling.Cqrs;
+using PlatformPlatform.Foundation.DomainModeling.Validation;
 
 namespace PlatformPlatform.AccountManagement.Application.Tenants.Commands;
 
@@ -46,7 +47,7 @@ public static class CreateTenant
 
                 RuleFor(x => x.Name).SetValidator(new TenantPropertyValidation.Name());
                 RuleFor(x => x.Email).SetValidator(new TenantPropertyValidation.Email());
-                RuleFor(x => x.Phone).SetValidator(new TenantPropertyValidation.Phone());
+                RuleFor(x => x.Phone).SetValidator(new SharedValidations.Phone());
                 RuleFor(x => x.Subdomain).SetValidator(new TenantPropertyValidation.Subdomain());
                 RuleFor(x => x.Subdomain)
                     .MustAsync(SubdomainMustBeAvailable)
